로그인추가해봦

+++ coupang/src/coupang/MemberMenu.java
@@ -167,7 +167,13 @@ public class MemberMenu {
 						String name = sc.next();
 						System.out.println("주민번호 : ");
 						String residentNumber = sc.next();
-						mc.deleteCustomer(name, residentNumber);
+						System.out.println("정말 삭제 하시겠습니까? (y/n)");
+						String str = sc.next();
+						if(str.equalsIgnoreCase("Y"))
+							mc.deleteCustomer(name, residentNumber);
+						else
+							System.out.println("삭제를 취소합니다");
+						
 					}
 					else 
 						System.out.println("잘 못 입력하셨습니다");
